Removed hardcoded path to sd card and changed the checkbox onclick to a String.
http://stackoverflow.com/questions/32594444/checkbox-append-text-to-string-if-checked-then-write-either-empty-string-or-appe

"I'm assuming that you have the CheckBoxes working as the code looks like will work. I will warn you that how you currently have it you will be appending the text every that you check and uncheck the checkBox as the append is never undone. I believe your current code was not working because you never actually called the print command for the FileOutStream so you added data but never printed it

For writing to the SD card couple of things that I would recommend(In code below). Instead of using the OutStreamWriter use a StringBuffer for appending the text to and a PrintStream to write to file as this allows to write Strings to the file. It's also poor practice to hardcode where the SD card is and you should use Environment.getExternalStorageDirectory() rather than "/sdcard""